Make course list handle all the logic

+++ src/app/components/course-list/course-list.component.ts
@@ -3,6 +3,8 @@ import { CourseServiceService } from '../../services/course-service.service';
 import Course from '../../models/Course'
 import { FilterbytextPipe } from '../../pipes/filterbytext.pipe';
 import { SearchtextService } from 'src/app/services/searchtext.service';
+import { MatDialog } from '@angular/material';
+import { DialogConfirmationComponent } from '../dialog-confirmation/dialog-confirmation.component';
 
 @Component({
   selector: 'app-course-list',
@@ -18,7 +20,8 @@ export class CourseListComponent implements OnInit {
   constructor(
     private courseService: CourseServiceService, 
     private filterByText: FilterbytextPipe,
-    private textService: SearchtextService) { 
+    private textService: SearchtextService,
+    public dialog: MatDialog) { 
     console.log('constructor')
   }
 
@@ -66,7 +69,17 @@ export class CourseListComponent implements OnInit {
 
   deleteCourse(id: number){
     console.log('Delete course with ID: ', id)
-    this.courseService.removeItem(id);
-    this.coursesFromService = this.courseService.getCourses();
+    this.openConfirmationDialog(id);
+  }
+
+  openConfirmationDialog(id: number){
+    let courseIndex = this.coursesToDisplay.findIndex((course)=>{return course.id === id})
+    this.dialog.open(DialogConfirmationComponent, { data: { title: this.coursesToDisplay[courseIndex].title} })
+      .afterClosed().subscribe(res =>{
+        if(res){
+          this.courseService.removeItem(id);
+          this.coursesFromService = this.courseService.getCourses();
+        }
+    });   
   }
 }
